move Player class to player package
