+++ src/com/yoyojin/edu/dice_game/rules/WinRule.java
@@ -1,7 +1,7 @@
 package com.yoyojin.edu.dice_game.rules;
 
 import com.yoyojin.edu.dice_game.exceptions.MoreThanOneWinnerException;
-import com.yoyojin.edu.dice_game.models.Player;
+import com.yoyojin.edu.dice_game.models.player.Player;
 
 import java.util.List;
